Structural importance + float32 storage win + measured perf benchmark
- StructuralImportance (no LLM): content salience + assertion/attribute structure — provably fixes
  a ContentImportance blind spot (a fact outranks a same-topic question), but NEUTRAL on the
  retention benchmark. Third importance signal to land neutral → the bottleneck is the eval, not
  the signal (recall@k on LoCoMo/multiday can't isolate importance). Off by default; 5 tests.
- float32 in-memory embeddings: records store the embedding as a float32 numpy array, not a
  list[float]. Measured ~7x smaller footprint (~24 -> ~3.6 MB/1k at 768 dims) and faster queries;
  recall@k unchanged (LoCoMo 0.27). SQLite already persisted float32.
- eval/bench_perf.py: latency / throughput / REAL tracemalloc footprint (no LLM) — the perf numbers
  the project never measured. README "millisecond operations" softened to honest "tens of ms,
  embed-bound". BENCHMARKS.md perf section added.

14 prior + 5 new tests green (102). See docs §7 + CHANGELOG + BENCHMARKS §2.

+++ midas/store.py
@@ -64,7 +64,7 @@ class InMemoryStore:
             return
         recs = [r for r in self._records.values() if r.embedding is not None]
         self._matrix_recs = recs
-        self._matrix = np.asarray([r.embedding for r in recs], dtype=np.float64) if recs else None
+        self._matrix = np.asarray([r.embedding for r in recs], dtype=np.float32) if recs else None
         self._dirty = False
 
     def search(
@@ -97,7 +97,7 @@ class InMemoryStore:
         if self._matrix is None:
             return []
         recs = self._matrix_recs
-        sims = np.clip(self._matrix @ np.asarray(embedding, dtype=np.float64), -1.0, 1.0)
+        sims = np.clip(self._matrix @ np.asarray(embedding, dtype=np.float32), -1.0, 1.0)
 
         if predicate is not None:
             mask = np.fromiter((predicate(r) for r in recs), dtype=bool, count=len(recs))
